requirements.txt
fastapi
uvicorn
langchain
langchain-groq
openai
snowflake-connector-python
pyyaml
python-dotenv
pandas